Move IFileType to interfaces directory
Relocated IFileType.ts to a new 'interfaces' directory to improve folder structure and organization. Updated import paths in FileType.ts and related unit tests accordingly to maintain proper functionality. This enhances code maintainability and clarity by grouping related interfaces together.
