+++ src/types/FileType/interfaces/IFileType.ts
